Rollback

+++ OpenGL Sandbox/Light.cpp
@@ -15,7 +15,7 @@ Light::Light()
     diffuseIntensity = 0.0f;
 }
 
-Light::Light(GLuint shadowWidth, GLuint shadowHeight, GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity)
+Light::Light(GLfloat shadowWidth, GLfloat shadowHeight, GLfloat red, GLfloat green, GLfloat blue, GLfloat aIntensity, GLfloat dIntensity)
 {
     shadowMap = new ShadowMap();
     shadowMap->Init(shadowWidth, shadowHeight);
